(refactor) Moves testing preliminaries to header file
